cleanup unnecessary casts, generics, missing .cvsignores, some comments.

+++ components/gnutella-core/src/main/java/com/limegroup/bittorrent/BTDownloader.java
@@ -282,7 +282,7 @@ implements TorrentEventListener {
 	}
 
 	public int getChunkSize() {
-		return (int) _info.getPieceLength();
+		return _info.getPieceLength();
 	}
 
 	public long getAmountLost() {
